requirements.txt
google-genai
google-cloud-firestore
google-auth
python-dotenv